+++ packages/antwerp-ui/react-components/src/lib/molecules/datepicker/Datepicker.tsx
@@ -165,22 +165,27 @@ export function Datepicker({
     }
   };
 
+  const inputId = inputProps?.id ?? 'aui-text-field';
+
   return (
     <div className="a-input has-icon-right" data-qa={qa}>
-      {(label || inputProps?.label) && renderLabel({ label: label || inputProps?.label, id: inputProps?.id, required })}
+      {(label || inputProps?.label) && renderLabel({ label: label || inputProps?.label, id: inputId, required })}
       {renderDescription({
+        id: inputId,
         description: dateInvalidError || inputProps?.description,
         state: dateInvalidError ? 'error' : inputProps?.state
       })}
       <div className="a-input__wrapper">
         <TextField
           {...inputProps}
+          id={inputId}
           label={undefined}
           description={undefined}
           type="text"
           value={inputProps.value || formattedValue}
           onChange={handleChange}
           state={dateInvalidError ? 'error' : inputProps?.state}
+          aria-describedby={`${inputId}--description`}
         />
         <Icon
           tabIndex={0}