+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed a margin bottom issue with the inline alert title
+- Fixed the `aria-describedby` attribute not being linked to `description` text in Datepicker component
 
 ## 7.11.1 - 2026-01-07